modified request method for userwallet and verification endpoints

+++ api/models.py
@@ -130,8 +130,8 @@ class Verification(models.Model):
 
 class UserWallet(models.Model):
 	user = models.ForeignKey(User, on_delete=models.CASCADE)
-	available_balance = models.FloatField(default=0.00)
-	total_balance = models.FloatField(default=0.00)
+	available_balance = models.DecimalField(default=0.00, max_digits=1000000, decimal_places=2)
+	total_balance = models.DecimalField(default=0.00, max_digits=1000000, decimal_places=2)
 
 	def __str__(self):
 		return str(self.available_balance)
